+++ discovery/repository/redshift/redshift_test.go
@@ -14,7 +14,7 @@ import (
 
 func TestListDatabases(t *testing.T) {
 	ctx, db, mock, r := initRepoTest(t)
-	defer db.Close()
+	defer func() { _ = db.Close() }()
 	dbRows := sqlmock.NewRows([]string{"name"}).AddRow("db1").AddRow("db2")
 	mock.ExpectQuery(postgresql.DatabaseQuery).WillReturnRows(dbRows)
 	dbs, err := r.ListDatabases(ctx)
@@ -22,11 +22,11 @@ func TestListDatabases(t *testing.T) {
 	require.ElementsMatch(t, []string{"db1", "db2"}, dbs)
 }
 
-func initRepoTest(t *testing.T) (context.Context, *sql.DB, sqlmock.Sqlmock, *redshiftRepository) {
+func initRepoTest(t *testing.T) (context.Context, *sql.DB, sqlmock.Sqlmock, *Repository) {
 	ctx := context.Background()
 	db, mock, err := sqlmock.New()
 	require.NoError(t, err)
-	return ctx, db, mock, &redshiftRepository{
-		genericSqlRepo: genericsql.NewGenericSqlRepositoryFromDB("repoName", RepoTypeRedshift, "dbName", db),
+	return ctx, db, mock, &Repository{
+		genericSqlRepo: genericsql.NewRepositoryFromDB("repoName", RepoTypeRedshift, "dbName", db),
 	}
 }
